moved CurrentLogwatchersMeterBinder into own maven and guice module and events into kubernetes module
